+++ showcase/starters/google-adk/agent/registry.py
@@ -24,10 +24,14 @@ from .shared_chat import (
 )
 from .gen_ui_agent import gen_ui_agent
 from .gen_ui_tool_based_agent import gen_ui_tool_based_agent
-from .tool_rendering_agents import (
-    tool_rendering_agent,
+from .tool_rendering_agent import tool_rendering_agent
+from .tool_rendering_default_catchall_agent import (
     tool_rendering_default_catchall_agent,
+)
+from .tool_rendering_custom_catchall_agent import (
     tool_rendering_custom_catchall_agent,
+)
+from .tool_rendering_reasoning_chain_agent import (
     tool_rendering_reasoning_chain_agent,
 )
 from .shared_state_read_write_agent import (
